refactor comments into dedicated Comment component
